Move generated code (Go) to the root dir
It will be used in both server (backend) and client parts in the future.
